add memoizeWith
